Update GetById to return nested data
The initial implementation of the repository layer was using lazy loading to return the SingleOrDefault. By converting it to a list we get the nested values of DirectReports, though at the cost of dumping the entire DbSet to a list. There is probably a better way to handle this, maybe linq includes or recursion. I acknowledge this is a tad bit hacky.