Add ability to get reporting structure
Straight forward approach that can use the existing controller (as it is based on the Employee). It feels like this may end up being a preview/detail view situation where we get a light amount of data if we're looking at a large set, and then can see the detail view when doing a singular get call.

I added the model and a dto as I wanted to extract the two. In this case I still return the model as part of the OkObjectResult, but the Dto handles the logic to calculate the numberOfReports.

I updated the integration tests to setup/teardown for each, as the modification test would change the state between tests and my added test would not be valid after the put request test, but is fine in isolation. This is a naive approach, and likely I would add the ability to reset state between tests as an option.

I added some unit tests, utilizing my commonly used framework to show the approach I would take with those versus needing a full working component.