Add create and get commands for compensation
Added new repository (with dbcontext), service, and controller layer as well as entity model and dto. Made some executive decisions based on the request, seems like more than one compensation could exist for a given employee, as the effectiveDate is the differentiator there, so I return a list of compensations when given an employee id. When creating, I have the controller use the employee service to get the details for said employee, as pasting that into the request would be daunting, but the returned data via the dto allows the request to still contain all the employee information. The dto has a collection variant on top of the singleton to handle the create response (or a single get if that's wanted later on) differently.